requirements.txt
# 核心依赖
numpy>=1.20.0
scipy>=1.7.0

# 配置管理
pyyaml>=5.4.1

# 文本提取
pdfminer.six>=20221105
ebooklib>=0.18
pypdf2>=3.0.0  # 备用PDF库

# OCR支持（可选）
# 需要先安装Tesseract OCR引擎: https://github.com/tesseract-ocr/tesseract
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0

# NLP处理
spacy>=3.4.0
# 下载Spacy模型: python -m spacy download en_core_web_sm

# 深度学习
torch>=1.12.0
transformers>=4.21.0

# 拓扑数据分析
ripser>=0.6.4

# 可视化
kmapper>=2.0.1
umap-learn>=0.5.3

# 机器学习工具
scikit-learn>=1.0.0

# 进度条
tqdm>=4.64.0

# 日志和工具
colorlog>=6.7.0  # 可选：彩色日志输出

